feat: refactor encoding stats code to be its' own file
